+++ src/CardboardBox.Anime.Database/Scripts/Manga/proc_get_manga_filtered.sql
@@ -1,11 +1,11 @@
-CREATE OR REPLACE FUNCTION get_manga_filtered(platformId text, state int, ids bigint[])
+CREATE OR REPLACE FUNCTION get_manga_filtered(platformId text, state integer, ids bigint[])
     RETURNS TABLE (
         manga_id BIGINT,
         manga_chapter_id BIGINT,
         first_chapter_id BIGINT,
         progress_chapter_id BIGINT,
         progress_id BIGINT,
-        max_chapter_num NUMERIC,
+        max_chapter_num INTEGER,
         chapter_num INTEGER,
         page_count INT,
         chapter_progress NUMERIC,
